+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 
 ## 2023 March
 
+March 31:
+- Change nggui from int to double.
+- Change test from int to double.
+
 March 28:
 - Change nggraphics from int to double.
 - Collider code for vec, rect, and mass.